+++ src/LiveComponent/src/Util/LiveControllerAttributesCreator.php
@@ -121,6 +121,21 @@ class LiveControllerAttributesCreator
             );
         }
 
+        $liveMetadata = $this->metadataFactory->getMetadata($mounted->getName());
+        if ($liveMetadata->hasQueryStringBindings()) {
+            $queryMapping = [];
+            foreach ($liveMetadata->getAllLivePropsMetadata() as $livePropMetadata) {
+                if ($mapping = $livePropMetadata->getQueryStringMapping()) {
+                    foreach ($mapping['parameters'] as $parameter => $config) {
+                        $queryMapping[$config['property']] = [
+                            'name' => $parameter,
+                        ];
+                    }
+                }
+            }
+            $attributesCollection->setQueryUrlMapping($queryMapping);
+        }
+
         return $attributesCollection;
     }